Java ATM System.
It is an ATM (Automated banking machine) interface using Java. In the process we will carry out 5 classes: for the account, its owner, the transactions, the bank and the ATM itself. It is used to access the services, the user will have to enter an identifier with a pin code. You will then have access to a statement of transactions, the ability to top up and withdraw money from the account, and make transfers.